sesion_07/requirements/local.txt
-r ./base.txt

django-extensions>=3.0.9,<3.1.0
